% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_cv.R
\name{corr_cv}
\alias{corr_cv}
\title{A Crossvalidated Correlation Function}
\usage{
corr_cv(x, y)
}
\arguments{
\item{x}{The first variable you want to correlate.}

\item{y}{The second variable you want to correlate.}
}
\description{
This function allows you to examine crossvalidated correlations for your data.
}
\examples{
corr_cv(x,y)
corr_cv()
}
\keyword{correlation}
\keyword{correlation,}
\keyword{crossvalidated}
