% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_error.R
\name{std_error}
\alias{std_error}
\title{A Standard Error Function}
\usage{
std_error(x)
}
\arguments{
\item{x}{The data you want to estimate standard error. Can be a vector or a matrix}
}
\description{
This function allows you to examine the standard error of the mean.
}
\examples{
std_error(x)
std_error()
}
\keyword{descriptives}
\keyword{error,}
\keyword{standard}
\keyword{variability,}
