% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.omit.R
\name{string.omit}
\alias{string.omit}
\title{A Function to Remove Columns with Strings}
\usage{
string.omit(ds)
}
\arguments{
\item{ds}{The dataframe from which you want to remove all variables that are characters}
}
\description{
This function allows you to remove all columns with class "character"
}
\examples{
string.omit(x)
string.omit()
}
\keyword{omit}
\keyword{omit,}
\keyword{string}
